git-pw: Filtering by test state is orthogonal to other filters
The way is was done in the initial submission is that all the filters
were exclusive, because they were also defining an ordering (so using
two filters would have conflicting ordering of the results).

I'd rather have ordering being orthogonal to the filters, but do include
some smarts about the default ordering if the filters include
timestamps.

As a first step, make sure we can filter by test-state *and* other
things.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>