series: Compute a state for series along with a summary patch statuses
We really want to be able to tell if we're finished with a series or
not. For instance this can be used to not display 'DONE' series in the
default list of series. It can also be used to filter out series done
with from the list of reviews a user has to do.

We take the opportunity to incorporate an 'incomplete' state, which
represents series with a revision that didn't receive all of its
patches.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>